requirements.txt
# Project W - Production Dependencies
# Generated from pyproject.toml

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0
structlog>=24.1.0
python-json-logger>=2.0.7
prometheus-client>=0.20.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.1
redis>=5.0.4
boto3>=1.34.131
pgvector>=0.2.5
tenacity>=8.3.0
orjson>=3.10.3
python-dotenv>=1.0.1
openai>=1.0.0
google-generativeai>=0.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
celery>=5.3.4
anthropic>=0.7.7
email-validator>=2.0.0
numpy>=1.24.0

